move route helpers to Managers folder
